# docs/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/contributions": {
      "post": {
        "operationId": "ContributionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContributionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "contributions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/contributions/mine": {
      "get": {
        "operationId": "ContributionsController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "contributions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/contributions/mine/{id}": {
      "get": {
        "operationId": "ContributionsController_getMineById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "contributions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ContributionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContributionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "contributions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/contributions/mine/{id}/withdraw": {
      "post": {
        "operationId": "ContributionsController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawContributionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "contributions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/contributions/{id}/provenance-sources": {
      "post": {
        "operationId": "ContributionsController_addProvenanceSource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProvenanceSourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "contributions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/contributions": {
      "get": {
        "operationId": "ContributionsAdminController_queue",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "SUBMITTED",
                "TRIAGE",
                "PROVENANCE_REVIEW",
                "HISTORICAL_REVIEW",
                "ACCEPTED",
                "CATALOGUED",
                "REJECTED"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DOCUMENT",
                "PHOTO",
                "ARCHIVAL_PHOTO",
                "MAP",
                "ORAL_HISTORY",
                "PERSONAL_MEMORY",
                "FAMILY_ARCHIVE",
                "BOOK_REFERENCE",
                "LOCAL_HISTORY",
                "CORRECTION",
                "OTHER"
              ],
              "type": "string"
            }
          },
          {
            "name": "sensitivity",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "NORMAL",
                "HIGH",
                "TERRITORIAL",
                "LEGAL",
                "CONTESTED"
              ],
              "type": "string"
            }
          },
          {
            "name": "submitterId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reviewerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "correctionTargetType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PLACE",
                "PERSON",
                "EVENT",
                "ERA",
                "DYNASTY",
                "TERRITORY",
                "FACT",
                "STORY",
                "JOURNEY",
                "COMMUNITY_STORY",
                "SOURCE",
                "SOURCE_DOCUMENT",
                "CONTRIBUTION",
                "MEDIA_ASSET",
                "COMMENT",
                "COUNTRY",
                "REGION",
                "CITY",
                "DESTINATION",
                "PROVIDER",
                "PROVIDER_LICENSE",
                "PROVIDER_INTEGRATION"
              ],
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin/contributions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/countries": {
      "get": {
        "operationId": "CountriesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "countries"
        ]
      },
      "post": {
        "operationId": "CountriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCountryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "countries"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/countries/{slug}": {
      "get": {
        "operationId": "CountriesController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "countries"
        ]
      }
    },
    "/v1/countries/{slug}/regions": {
      "get": {
        "operationId": "CountriesController_getRegions",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "countries"
        ]
      }
    },
    "/v1/countries/{slug}/cities": {
      "get": {
        "operationId": "CountriesController_getCities",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "region",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "countries"
        ]
      }
    },
    "/v1/countries/{slug}/destinations": {
      "get": {
        "operationId": "CountriesController_getDestinations",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "region",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "countries"
        ]
      }
    },
    "/v1/countries/{id}": {
      "patch": {
        "operationId": "CountriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCountryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "countries"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/countries/{id}/translations/{locale}": {
      "patch": {
        "operationId": "CountriesController_upsertTranslation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertCountryTranslationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "countries"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/countries/{id}/status": {
      "patch": {
        "operationId": "CountriesController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCountryStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "countries"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Dau Viet API",
    "description": "Living Digital Atlas of Vietnam - backend API contract for web, mobile, and admin clients. See docs/backend/BACKEND_HANDOFF.md for the full handoff contract this document is generated alongside.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateContributionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "DOCUMENT",
              "PHOTO",
              "ARCHIVAL_PHOTO",
              "MAP",
              "ORAL_HISTORY",
              "PERSONAL_MEMORY",
              "FAMILY_ARCHIVE",
              "BOOK_REFERENCE",
              "LOCAL_HISTORY",
              "CORRECTION",
              "OTHER"
            ],
            "default": "OTHER"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "originalLocale": {
            "type": "string",
            "description": "BCP-47-ish locale of the original submission text (spec section 7) - never silently translated."
          },
          "originSource": {
            "type": "string"
          },
          "currentOwner": {
            "type": "string"
          },
          "sharingRights": {
            "type": "string"
          },
          "submitterDeclaration": {
            "type": "string",
            "enum": [
              "OWN_MATERIAL",
              "HAVE_PERMISSION",
              "PUBLICLY_AVAILABLE",
              "UNKNOWN"
            ],
            "description": "Submitter's own claim - never an adjudicated rights state (spec section 13)."
          },
          "attribution": {
            "type": "string",
            "enum": [
              "NAMED",
              "ANONYMOUS",
              "INSTITUTIONAL"
            ]
          },
          "approxDateLabel": {
            "type": "string"
          },
          "approxDateStart": {
            "type": "string"
          },
          "approxDateEnd": {
            "type": "string"
          },
          "peopleShown": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "placeId": {
            "type": "string"
          },
          "linkedEntityType": {
            "type": "string",
            "enum": [
              "PERSON",
              "EVENT",
              "ERA",
              "TERRITORY"
            ]
          },
          "linkedEntityId": {
            "type": "string"
          },
          "correctionTargetType": {
            "type": "string",
            "enum": [
              "PLACE",
              "PERSON",
              "EVENT",
              "ERA",
              "STORY",
              "SOURCE",
              "FACT"
            ],
            "description": "Required (with correctionTargetId) when type=CORRECTION."
          },
          "correctionTargetId": {
            "type": "string"
          },
          "contextNote": {
            "type": "string"
          },
          "mediaAssetIds": {
            "description": "MediaAsset ids uploaded beforehand - each must be owned by the caller (or the caller is EDITOR+).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateContributionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "originSource": {
            "type": "string"
          },
          "currentOwner": {
            "type": "string"
          },
          "sharingRights": {
            "type": "string"
          },
          "submitterDeclaration": {
            "type": "string",
            "enum": [
              "OWN_MATERIAL",
              "HAVE_PERMISSION",
              "PUBLICLY_AVAILABLE",
              "UNKNOWN"
            ]
          },
          "attribution": {
            "type": "string",
            "enum": [
              "NAMED",
              "ANONYMOUS",
              "INSTITUTIONAL"
            ]
          },
          "approxDateLabel": {
            "type": "string"
          },
          "approxDateStart": {
            "type": "string"
          },
          "approxDateEnd": {
            "type": "string"
          },
          "peopleShown": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "contextNote": {
            "type": "string"
          },
          "mediaAssetIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "WithdrawContributionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "AddProvenanceSourceDto": {
        "type": "object",
        "properties": {
          "referenceType": {
            "type": "string",
            "enum": [
              "BOOK",
              "ARCHIVE",
              "WEBSITE",
              "FAMILY_RECORD",
              "ORIGINAL_DOCUMENT",
              "ORAL_TESTIMONY",
              "INSTITUTIONAL",
              "OTHER"
            ]
          },
          "claimedCreator": {
            "type": "string"
          },
          "claimedOwner": {
            "type": "string"
          },
          "acquisitionMethod": {
            "type": "string"
          },
          "approxDateLabel": {
            "type": "string"
          },
          "sourceOrganization": {
            "type": "string"
          },
          "archiveCatalogRef": {
            "type": "string"
          },
          "publicationInfo": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "evidenceMediaAssetId": {
            "type": "string"
          }
        },
        "required": [
          "referenceType"
        ]
      },
...
      "CountryTranslationInputDto": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "example": "vi"
          },
          "name": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "ORIGINAL",
              "HUMAN",
              "AI_ASSISTED"
            ]
          }
        },
        "required": [
          "locale",
          "name"
        ]
      },
      "CreateCountryDto": {
        "type": "object",
        "properties": {
          "iso2": {
            "type": "string",
            "example": "VN",
            "description": "ISO 3166-1 alpha-2"
          },
          "iso3": {
            "type": "string",
            "example": "VNM",
            "description": "ISO 3166-1 alpha-3"
          },
          "defaultLocale": {
            "type": "string",
            "example": "vi",
            "description": "The country's own default locale - a geography default, never a user preference."
          },
          "defaultCurrency": {
            "type": "string",
            "example": "VND",
            "description": "ISO 4217 currency code"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CountryTranslationInputDto"
            }
          }
        },
        "required": [
          "iso2",
          "iso3",
          "defaultLocale",
          "defaultCurrency",
          "translations"
        ]
      },
      "UpdateCountryDto": {
        "type": "object",
        "properties": {
          "defaultLocale": {
            "type": "string"
          },
          "defaultCurrency": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        }
      },
      "UpsertCountryTranslationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "ORIGINAL",
              "HUMAN",
              "AI_ASSISTED"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "SetCountryStatusDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}